extract handleAliasType into its own module
